optimization & refactor: optimized button component, refactor header items and secondary button in modals
