+++ src/components/Buttons/LinkedButton.tsx
@@ -3,9 +3,9 @@ import {LinkedButtonProps} from "@/components/Buttons/types";
 import Link from "next/link";
 import styles from "./Button.module.scss";
 import {useGenerateId} from "@/hooks/useGenerateId";
-import btnTypes from "@/components/Buttons/buttonTypes.module.scss";
-import btnRadius from './buttonRadius.module.scss';
-import btnSizes from './buttonSizes.module.scss';
+import btnTypes from "@/components/Buttons/types.module.scss";
+import btnRadius from './radius.module.scss';
+import btnSizes from './sizes.module.scss';
 
 const LinkedButton = forwardRef<HTMLAnchorElement, LinkedButtonProps>((
 	{
@@ -20,6 +20,8 @@ const LinkedButton = forwardRef<HTMLAnchorElement, LinkedButtonProps>((
 		target = '_self',
 		radius = 'medium',
 		size = 'full',
+		onClick,
+		children,
   }, ref) => {
 		const id = useGenerateId();
 		return (
@@ -33,10 +35,13 @@ const LinkedButton = forwardRef<HTMLAnchorElement, LinkedButtonProps>((
 				aria-labelledby={ariaLabelledBy}
 				ref={ref}
 				target={target}
+				onClick={onClick}
+				draggable="false"
 			>
         <span>
           {Icon && <Icon className={styles.icon} />}
 	        {text && text}
+					{children && children}
         </span>
 			</Link>
 		);
